+++ templates/_inc/_header.html.twig
@@ -19,6 +19,10 @@
                         <li class="nav-item">
                             <a class="nav-link" href="{{ path('admin') }}">Admin</a>
                         </li>
+                    {% else %}
+                        <li class="nav-item">
+                            <a class="nav-link" href="{{ path('user') }}">User</a>
+                        </li>
                     {% endif %}
                     <li class="nav-item">
                         <a class="nav-link" href="{{ path('app_logout') }}">Logout</a>
