--- templates/comment/_form.html.twig
@@ -1,10 +1,0 @@
-{{ form_start(form) }}
-    <div class="row">
-        <div class="col-12 col-md-4 offset-md-3 text-center">
-            {{ form_widget(form) }}
-        </div>
-        <div class="col-12 col-md-2 text-center">
-            <button class="btn btn-primary">{{ button_label|default('Save') }}</button>
-        </div>
-    </div>
-{{ form_end(form) }}
