+++ templates/common/image/_form.html.twig
