+++ templates/common/new.html.twig
@@ -6,6 +6,6 @@
     <div class="container">
         <h1>Create new {{ name|capitalize }}</h1>
 
-        {{ include(name ~ '/_form.html.twig') }}
+        {{ include('common/' ~ name ~ '/_form.html.twig') }}
     </div>
 {% endblock %}
