+++ templates/front/register.html.twig
