--- templates/image/show.html.twig
@@ -1,32 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Image{% endblock %}
-
-{% block body %}
-    <h1>Image</h1>
-
-    <table class="table">
-        <tbody>
-            <tr>
-                <th>Id</th>
-                <td>{{ image.id }}</td>
-            </tr>
-            <tr>
-                <th>Name</th>
-                <td>
-                    <img src="{{ asset('uploads/images/' ~ image.name) }}" alt="{{ image.alt }}" style="width: 300px">
-                </td>
-            </tr>
-            <tr>
-                <th>Alt</th>
-                <td>{{ image.alt }}</td>
-            </tr>
-        </tbody>
-    </table>
-
-    <a href="{{ path('app_image_index') }}">back to list</a>
-
-    <a href="{{ path('app_image_edit', {'id': image.id}) }}">edit</a>
-
-    {{ include('image/_delete_form.html.twig') }}
-{% endblock %}
