--- templates/trick/show.html.twig
@@ -1,63 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Trick{% endblock %}
-
-{% block body %}
-    <div class="container">
-        <div class="trick-main-image">
-            <h1>{{ trick.name }}</h1>
-            <img src="{{ asset('images/snowtricks.jpg') }}" alt="SnowTricks" class="img-fluid">
-        </div>
-        <div class="row">
-
-        </div>
-        <div class="row">
-            <div class="col-md-10 offset-md-1">
-                <p>{{ trick.description }}</p>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-12 col-md-6 offset-md-3 text-center">
-                <span class="badge text-bg-primary">Primary</span>
-                <span class="badge text-bg-secondary">Secondary</span>
-                <span class="badge text-bg-success">Success</span>
-                <span class="badge text-bg-danger">Danger</span>
-
-                <hr>
-
-            </div>
-
-            {% if is_granted('ROLE_USER') %}
-                {{ include('comment/_form.html.twig', {'button_label': 'Leave a comment'}) }}
-            {% endif %}
-
-            <div class="col-12 col-md-6 offset-md-3">
-                <hr>
-
-                <h2>Comments</h2>
-                <div class="accordion" id="accordionExample">
-                    {% for key, comment in trick.comments %}
-                        <div class="accordion-item">
-                            <h2 class="accordion-header">
-                                <button class="accordion-button" type="button" data-bs-toggle="collapse" data-bs-target="#collapse-{{ comment.id }}" aria-expanded="{% if key is same as (0) %}true{% else %}false{% endif %}" aria-controls="collapse-{{ comment.id }}">
-                                    <img class="img-thumbnail me-2" src="{{ asset('images/snowtricks.jpg') }}" style="width: 50px">
-                                    Accordion Item #{{ comment.id }}
-                                </button>
-                            </h2>
-                            <div id="collapse-{{ comment.id }}" class="accordion-collapse collapse {% if key is same as (0) %}show{% endif %}" data-bs-parent="#accordionExample">
-                                <div class="accordion-body">
-                                    <strong>This is the first item's accordion body.</strong> It is shown by default, until the collapse plugin adds the appropriate classes that we use to style each element. These classes control the overall appearance, as well as the showing and hiding via CSS transitions. You can modify any of this with custom CSS or overriding our default variables. It's also worth noting that just about any HTML can go within the <code>.accordion-body</code>, though the transition does limit overflow.
-                                </div>
-                            </div>
-                        </div>
-                    {% else %}
-                        <p>No comments yet.</p>
-                    {% endfor %}
-                </div>
-                <div class="text-center">
-                    <button type="button" class="btn btn-primary">Load more</button>
-                </div>
-            </div>
-        </div>
-    </div>
-{% endblock %}
