+++ templates/user/index.html.twig
@@ -3,18 +3,10 @@
 {% block title %}Administration{% endblock %}
 
 {% block body %}
-    <h1 class="text-center">Administration</h1>
-
     <div class="container">
         <ul class="nav nav-tabs" id="myTab" role="tablist">
             <li class="nav-item" role="presentation">
                 <button class="nav-link active" id="trick-tab" data-bs-toggle="tab" data-bs-target="#trick-tab-pane" type="button" role="tab" aria-controls="trick-tab-pane" aria-selected="true">Tricks</button>
-            </li>
-            <li class="nav-item" role="presentation">
-                <button class="nav-link" id="category-tab" data-bs-toggle="tab" data-bs-target="#category-tab-pane" type="button" role="tab" aria-controls="category-tab-pane" aria-selected="false">Categories</button>
-            </li>
-            <li class="nav-item" role="presentation">
-                <button class="nav-link" id="user-tab" data-bs-toggle="tab" data-bs-target="#user-tab-pane" type="button" role="tab" aria-controls="user-tab-pane" aria-selected="false">Users</button>
             </li>
             <li class="nav-item" role="presentation">
                 <button class="nav-link" id="comment-tab" data-bs-toggle="tab" data-bs-target="#comment-tab-pane" type="button" role="tab" aria-controls="comment-tab-pane" aria-selected="false">Comments</button>
@@ -30,16 +22,6 @@
             <div class="tab-pane fade show active" id="trick-tab-pane" role="tabpanel" aria-labelledby="trick-tab" tabindex="0">
                 {{ include('_inc/_user/_trick.html.twig', {
                     tricks: tricks,
-                }) }}
-            </div>
-            <div class="tab-pane fade" id="category-tab-pane" role="tabpanel" aria-labelledby="category-tab" tabindex="0">
-                {{ include('_inc/_admin/_category.html.twig', {
-                    categories: categories,
-                }) }}
-            </div>
-            <div class="tab-pane fade" id="user-tab-pane" role="tabpanel" aria-labelledby="user-tab" tabindex="0">
-                {{ include('_inc/_admin/_user.html.twig', {
-                    users: users,
                 }) }}
             </div>
             <div class="tab-pane fade" id="comment-tab-pane" role="tabpanel" aria-labelledby="comment-tab" tabindex="0">
