--- templates/video/show.html.twig
@@ -1,26 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Video{% endblock %}
-
-{% block body %}
-    <h1>Video</h1>
-
-    <table class="table">
-        <tbody>
-            <tr>
-                <th>Id</th>
-                <td>{{ video.id }}</td>
-            </tr>
-            <tr>
-                <th>Url</th>
-                <td>{{ video.url }}</td>
-            </tr>
-        </tbody>
-    </table>
-
-    <a href="{{ path('app_video_index') }}">back to list</a>
-
-    <a href="{{ path('app_video_edit', {'id': video.id}) }}">edit</a>
-
-    {{ include('video/_delete_form.html.twig') }}
-{% endblock %}
