+++ tests/Controller/User/ImageControllerTest.php
@@ -9,37 +9,10 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class ImageControllerTest extends BaseController
 {
-    public function testImageHomePageWithoutLogin(): void
-    {
-        $client = static::createClient();
-        $client->request('GET', '/admin/image/');
-
-        $this->assertResponseStatusCodeSame(302);
-        $this->assertResponseRedirects('/login');
-    }
-
-    public function testImageHomePageWithUserLogin(): void
-    {
-        $client = $this->loginUser();
-        $client->request('GET', '/admin/image/');
-
-        $this->assertResponseStatusCodeSame(403);
-        $this->assertSelectorTextContains('h1.exception-message', 'Access Denied.');
-    }
-
-    public function testImageHomePageWithAdminLogin(): void
-    {
-        $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/');
-
-        $this->assertResponseStatusCodeSame(200);
-        $this->assertPageTitleContains('Image index');
-    }
-
     public function testImageNewPageWithoutLogin(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/admin/image/new');
+        $client->request('GET', '/user/image/new');
 
         $this->assertResponseRedirects('/login');
     }
@@ -47,16 +20,15 @@ class ImageControllerTest extends BaseController
     public function testImageNewPageWithUserLogin(): void
     {
         $client = $this->loginUser();
-        $client->request('GET', '/admin/image/new');
+        $client->request('GET', '/user/image/new');
 
-        $this->assertResponseStatusCodeSame(403);
-        $this->assertSelectorTextContains('h1.exception-message', 'Access Denied.');
+        $this->assertResponseStatusCodeSame(200);
     }
 
     public function testImageNewPageWithAdminLoginWithoutFormSubmission(): void
     {
         $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/new');
+        $client->request('GET', '/user/image/new');
 
         $this->assertResponseStatusCodeSame(200);
         $this->assertPageTitleContains('New Image');
@@ -65,78 +37,50 @@ class ImageControllerTest extends BaseController
     public function testImageNewPageWithAdminLoginWithFormSubmissionWithoutValues(): void
     {
         $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/new');
+        $client->request('GET', '/user/image/new');
         $client->submitForm('Save');
 
         $this->assertResponseStatusCodeSame(422);
         $this->assertPageTitleContains('New Image');
-        $this->assertSelectorTextContains('li', 'This value should not be blank.');
+        $this->assertSelectorTextContains('.invalid-feedback', 'This value should not be blank.');
     }
 
     public function testImageNewPageWithAdminLoginWithFormSubmissionWithInvalidValues(): void
     {
         $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/new');
+        $client->request('GET', '/user/image/new');
         $client->submitForm('Save', [
             'image[name]' => new UploadedFile(
-                __DIR__.'/../../public/uploads/images/snowtricks.txt',
+                __DIR__.'/../../../public/uploads/images/snowtricks.txt',
                 'snowtricks.txt'
             ),
         ]);
 
         $this->assertResponseStatusCodeSame(422);
         $this->assertPageTitleContains('New Image');
-        $this->assertSelectorTextContains('li', 'Please upload a valid image');
+        $this->assertSelectorTextContains('.invalid-feedback', 'Please upload a valid image');
     }
 
     public function testImageNewPageWithAdminLoginWithFormSubmissionWithValidValues(): void
     {
         $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/new');
+        $client->request('GET', '/user/image/new');
         $client->submitForm('Save', [
             'image[name]' => new UploadedFile(
-                __DIR__.'/../../public/uploads/images/snowtricks.jpg',
+                __DIR__.'/../../../public/uploads/images/snowtricks.jpg',
                 'snowtricks.jpg'
             ),
             'image[alt]' => 'Testing',
         ]);
 
         $this->assertResponseStatusCodeSame(303);
-        $this->assertResponseRedirects('/admin/image/');
-        $this->assertPageTitleContains('Redirecting to /admin/image/');
-    }
-
-    public function testImageShowPageWithoutLogin(): void
-    {
-        $client = static::createClient();
-        $client->request('GET', '/admin/image/1');
-
-        $this->assertResponseRedirects('/login');
-    }
-
-    public function testImageShowPageWithUserLogin(): void
-    {
-        $client = $this->loginUser();
-        $client->request('GET', '/admin/image/1');
-
-        $this->assertResponseStatusCodeSame(403);
-        $this->assertSelectorTextContains('h1.exception-message', 'Access Denied.');
-    }
-
-    public function testImageShowPageWithAdminLogin(): void
-    {
-        $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/1');
-
-        $this->assertResponseStatusCodeSame(200);
-        $this->assertPageTitleContains('Image');
-        $this->assertSelectorTextContains('h1', 'Image');
+        $this->assertResponseRedirects('/admin/');
     }
 
     public function testImageEditPageWithoutLogin(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/admin/image/1/edit');
+        $client->request('GET', '/user/image/1/edit');
 
         $this->assertResponseRedirects('/login');
     }
@@ -144,16 +88,15 @@ class ImageControllerTest extends BaseController
     public function testImageEditPageWithUserLogin(): void
     {
         $client = $this->loginUser();
-        $client->request('GET', '/admin/image/1/edit');
+        $client->request('GET', '/user/image/1/edit');
 
-        $this->assertResponseStatusCodeSame(403);
-        $this->assertSelectorTextContains('h1.exception-message', 'Access Denied.');
+        $this->assertResponseStatusCodeSame(200);
     }
 
     public function testImageEditPageWithAdminLoginWithUnexistingImage(): void
     {
         $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/100/edit');
+        $client->request('GET', '/user/image/100/edit');
 
         $this->assertResponseStatusCodeSame(404);
     }
@@ -161,23 +104,22 @@ class ImageControllerTest extends BaseController
     public function testImageEditPageWithAdminLoginWithExistingImage(): void
     {
         $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/1/edit');
+        $client->request('GET', '/user/image/1/edit');
         $client->submitForm('Update', [
             'image[name]' => new UploadedFile(
-                __DIR__.'/../../public/uploads/images/snowtricks-2.jpg',
+                __DIR__.'/../../../public/uploads/images/snowtricks-2.jpg',
                 'snowtricks-2.jpg'
             ),
         ]);
 
         $this->assertResponseStatusCodeSame(303);
-        $this->assertResponseRedirects('/admin/image/');
-        $this->assertPageTitleContains('Redirecting to /admin/image/');
+        $this->assertResponseRedirects('/admin/');
     }
 
     public function testImageRemovePageWithoutLogin(): void
     {
         $client = static::createClient();
-        $client->request('GET', '/admin/image/1');
+        $client->request('POST', '/user/image/1');
 
         $this->assertResponseRedirects('/login');
     }
@@ -185,39 +127,34 @@ class ImageControllerTest extends BaseController
     public function testImageRemovePageWithUserLogin(): void
     {
         $client = $this->loginUser();
-        $client->request('GET', '/admin/image/1');
+        $client->request('POST', '/user/image/1');
 
-        $this->assertResponseStatusCodeSame(403);
-        $this->assertSelectorTextContains('h1.exception-message', 'Access Denied.');
+        $this->assertResponseStatusCodeSame(303);
     }
 
     public function testImageRemovePageWithAdminLoginWithUnexistingImage(): void
     {
         $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/100');
+        $client->request('POST', '/user/image/100');
 
         $this->assertResponseStatusCodeSame(404);
     }
 
     public function testImageRemovePageWithAdminLoginWithExistingImageWithoutAssociation(): void
     {
-        $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/3');
-        $client->submitForm('Delete');
+        $client = $this->loginUser();
+        $client->request('POST', '/user/image/3');
 
         $this->assertResponseStatusCodeSame(303);
-        $this->assertResponseRedirects('/admin/image/');
-        $this->assertPageTitleContains('Redirecting to /admin/image/');
+        $this->assertResponseRedirects('/user/');
     }
 
     public function testImageRemovePageWithAdminLoginWithExistingImageWithAssociation(): void
     {
         $client = $this->loginAdmin();
-        $client->request('GET', '/admin/image/1');
-        $client->submitForm('Delete');
+        $client->request('POST', '/user/image/1');
 
         $this->assertResponseStatusCodeSame(303);
-        $this->assertResponseRedirects('/admin/image/');
-        $this->assertPageTitleContains('Redirecting to /admin/image/');
+        $this->assertResponseRedirects('/admin/');
     }
 }
